[FLINK-25432] Adds ResourceCleaner (+ default implementation) and integrates it into Dispatcher
Additionally, testHABlobsAreNotRemovedIfHAJobGraphRemovalFails is deleted
because this case doesn't apply anymore. The job graph deletion and the
HA file deletion happen independently from each other now: The JobResult
is already added to the JobResultStore as a dirty entry and all data can
be deleted.

There is a dependency between cleaning up the JobManagerRegistry (i.e.
closing the JobMaster) and cleaning up the HighAvailabilityServices for
the job. This issue is solved by a JobManager-wide leader election
(FLINK-24038). But we have to consider closing the JobMaster before
cleaning the HA job data until the legacy leader election functionality
is removed entirely.