[FLINK-29234] JobMasterServiceLeadershipRunner handle leader event in a separate executor to avoid dead lock.
Currently, in JobMasterServiceLeadershipRunner, Leader event callback such as grantLeadership and revokeLeadership will be invoked with nested lock(e.g. curator's event lock and DefaultLeaderElectionService's lock) and try to acquire JobMasterServiceLeadershipRunner's own lock. This will lead to many deadlock risks, such as the simultaneous close and grantLeadership. To avoid this, we should execute this call back in a separate thread to get rid of the nested structure of locks.