[review][hotfix] Adds JavaDoc after checking that the ExecutionAttemptID is actually correct
I double-checked that the JobInitializationMetricsBuilder constructor is called for every local restart of tasks and the JobInitializationMetricsBuilder#reportInitializationMetrics is called by the deployed Subtask after restoring its state. ExecutionAttemptID is correct because the builder creation happens again when a new Execution attempt is triggered (through the local restart).

The totalNumberOfSubtasks was a bit misleading here.